rename files to snake case
