+++ test/unit/client-side-encryption/auto_encrypter.test.ts
@@ -3,11 +3,11 @@ import * as fs from 'fs';
 import * as sinon from 'sinon';
 
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { AutoEncrypter } from '../../../src/client-side-encryption/autoEncrypter';
+import { AutoEncrypter } from '../../../src/client-side-encryption/auto_encrypter';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { MongocryptdManager } from '../../../src/client-side-encryption/mongocryptdManager';
+import { MongocryptdManager } from '../../../src/client-side-encryption/mongocryptd_manager';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { StateMachine } from '../../../src/client-side-encryption/stateMachine';
+import { StateMachine } from '../../../src/client-side-encryption/state_machine';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import { MongoClient } from '../../../src/mongo_client';
 import { BSON } from '../../mongodb';
