+++ test/unit/client-side-encryption/client_encryption.test.ts
@@ -4,16 +4,16 @@ import { resolve } from 'path';
 import * as sinon from 'sinon';
 
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { ClientEncryption } from '../../../src/client-side-encryption/clientEncryption';
+import { ClientEncryption } from '../../../src/client-side-encryption/client_encryption';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import * as cryptoCallbacks from '../../../src/client-side-encryption/cryptoCallbacks';
+import * as cryptoCallbacks from '../../../src/client-side-encryption/crypto_callbacks';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import {
   MongoCryptCreateDataKeyError,
   MongoCryptCreateEncryptedCollectionError
 } from '../../../src/client-side-encryption/errors';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { StateMachine } from '../../../src/client-side-encryption/stateMachine';
+import { StateMachine } from '../../../src/client-side-encryption/state_machine';
 import { Binary, BSON, deserialize } from '../../mongodb';
 
 const { EJSON } = BSON;
